Refatorando contratos do customer e mais algumas coisinhas

+++ src/main/java/com/jm/jamesapp/services/CustomerService.java
@@ -2,18 +2,18 @@ package com.jm.jamesapp.services;
 
 import com.jm.jamesapp.models.CustomerModel;
 import com.jm.jamesapp.models.UserModel;
+import com.jm.jamesapp.models.dto.SaveCustomerDto;
+import com.jm.jamesapp.models.dto.UpdateCustomerDto;
 import com.jm.jamesapp.repositories.CustomerRepository;
+import com.jm.jamesapp.services.exceptions.BusinessException;
 import com.jm.jamesapp.services.interfaces.ICustomerService;
+import com.jm.jamesapp.services.interfaces.ITransactionService;
 import com.jm.jamesapp.utils.constraints.CpfOrCnpjValidator;
-import com.jm.jamesapp.utils.constraints.ValidCpfOrCnpj;
-import jakarta.transaction.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
-import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 import static com.jm.jamesapp.utils.constraints.CpfOrCnpjValidator.cleanStringValue;
@@ -28,58 +28,83 @@ public class CustomerService implements ICustomerService {
     }
 
     @Override
-    @Transactional
-    public CustomerModel save(CustomerModel customerModel) {
-        customerModel.setBalance(BigDecimal.valueOf(0.0));
-        customerModel.setCpfCnpj(cleanStringValue(customerModel.getCpfCnpj()));
+    public CustomerModel save(SaveCustomerDto saveCustomerDto, UserModel userModel) {
+        saveCustomerDto.setCpfCnpj(cleanStringValue(saveCustomerDto.getCpfCnpj()));
 
-        customerRepository.save(customerModel);
-        return customerModel;
+        validateSave(saveCustomerDto, userModel);
+
+        CustomerModel customer = new CustomerModel(userModel, saveCustomerDto.getName(), saveCustomerDto.getCpfCnpj());
+
+        return customerRepository.save(customer);
     }
 
     @Override
-    public CustomerModel update(CustomerModel customerModel) {
-        customerModel.setCpfCnpj(cleanStringValue(customerModel.getCpfCnpj()));
-        customerRepository.save(customerModel);
-        return customerModel;
+    public CustomerModel update(CustomerModel customer, UpdateCustomerDto saveCustomerDto, UserModel userModel) {
+        validateUpdate(customer, saveCustomerDto, userModel);
+
+        customer.setName(saveCustomerDto.getName());
+        customer.setCpfCnpj(cleanStringValue(saveCustomerDto.getCpfCnpj()));
+
+        return customerRepository.save(customer);
     }
 
     @Override
-    public Page<CustomerModel> findAll(Pageable pageable) {
-        return customerRepository.findAll(pageable);
+    public CustomerModel findById(UUID id) {
+        return customerRepository.findById(id).orElse(null);
     }
 
     @Override
-    public Optional<CustomerModel> findById(UUID id) {
-        return customerRepository.findById(id);
-    }
-
-    @Override
-    @Transactional
+    // TODO: mas se o guerreiro quiser excluir mesmo assim?
+    // TODO: transactionService.calculateCustomerBalance(customerModel) ... valida se tem saldo
     public void delete(CustomerModel customerModel) {
-        // Regras para o Customers com balance > 0
+        BigDecimal balanceFromCustomer = calculateBalance(customerModel);
+        if (balanceFromCustomer.compareTo(BigDecimal.ZERO) > 0) throw new BusinessException("Cliente possui saldo pendente");
+
         customerRepository.delete(customerModel);
     }
 
-    @Override
-    public Optional<CustomerModel> findByIdAndOwner(UUID id, UserModel userModel) {
-        return customerRepository.findByIdAndOwner(id, userModel);
+    public BigDecimal calculateBalance(CustomerModel customerModel) {
+        return customerRepository.sumTransactionsByCustomer(customerModel);
+    }
+
+    public void deleteWithPendingBalance(CustomerModel customer){
+        customerRepository.delete(customer);
     }
 
     @Override
-    public Optional<CustomerModel> findByCpfCnpj(String cpfCnpj) {
-        cpfCnpj = CpfOrCnpjValidator.cleanStringValue(cpfCnpj);
-        return customerRepository.findByCpfCnpj(cpfCnpj);
+    public CustomerModel findByIdAndUser(UUID id, UserModel userModel) {
+        return customerRepository.findByIdAndUser(id, userModel).orElse(null);
     }
 
     @Override
-    public Optional<CustomerModel> findByCpfCnpjAndOwner(String cpfCnpj, UserModel userModel) {
-        cpfCnpj = CpfOrCnpjValidator.cleanStringValue(cpfCnpj);
-        return customerRepository.findByCpfCnpjAndOwner(cpfCnpj, userModel);
+    public Page<CustomerModel> findAllByUser(Pageable pageable, UserModel userModel) {
+        return customerRepository.findAllByUser(pageable, userModel);
     }
 
     @Override
-    public List<CustomerModel> findAllByOwner(UserModel userModel) {
-        return customerRepository.findAllByOwner(userModel);
+    public CustomerModel findByCpfCnpjAndUser(String cpfCnpj, UserModel userModel) {
+        return customerRepository.findByCpfCnpjAndUser(CpfOrCnpjValidator.cleanStringValue(cpfCnpj), userModel).orElse(null);
+    }
+
+    private void validateSave(SaveCustomerDto customerModel, UserModel userModel) {
+        boolean isCustomerAlreadyRegistered = findByCpfCnpjAndUser(customerModel.getCpfCnpj(), userModel) != null;
+        if (isCustomerAlreadyRegistered)
+            throw new BusinessException("Você possui um cliente cadastrado com o CPF/CNPJ informado.");
+    }
+
+    private void validateUpdate(CustomerModel customer, UpdateCustomerDto customerDto, UserModel userModel) {
+        if (customer == null || customerDto == null) throw new RuntimeException("");
+
+        var oldCpfCnpj = customer.getCpfCnpj();
+        var newCpfCnpj = cleanStringValue(customerDto.getCpfCnpj());
+
+        if (oldCpfCnpj.equals(newCpfCnpj) && customer.getName().equals(customerDto.getName()))
+            throw new BusinessException("Os novos dados são iguais aos dados originais. Nenhuma alteração foi feita.");
+
+        if (oldCpfCnpj.equals(newCpfCnpj)) return;
+
+        boolean isCustomerAlreadyRegistered = findByCpfCnpjAndUser(newCpfCnpj, userModel) != null;
+        if (!isCustomerAlreadyRegistered)
+            throw new BusinessException("Você possui um cliente cadastrado com o CPF/CNPJ informado.");
     }
 }
